Refactor Stubs
